fix eslint warning
